pixelize: avoid integer overflow when choosing code paths
For block sizes larger than 65535 the wrong code paths get chosen
since the squared size is larger than 32bit. Casting the values
compared to longs fixes issue #78 - https://gitlab.gnome.org/GNOME/gegl/issues/78

+++ operations/common/pixelize.c
@@ -569,7 +569,7 @@ process (GeglOperation       *operation,
     if (cl_process (operation, input, output, roi))
       return TRUE;
 
-  if (o->size_x * o->size_y < SQR (ALLOC_THRESHOLD_SIZE))
+  if ((long)o->size_x * (long)o->size_y < SQR (ALLOC_THRESHOLD_SIZE))
     {
       gfloat  background_color[4];
       gfloat *input_buf  = g_new (gfloat,
